[td] Don’t raise errors when adding dbt files (#4191)
* [td] Don’t raise errors when adding dbt files

* fix

* fix

+++ mage_ai/data_preparation/models/block/dbt/profiles.py
@@ -135,18 +135,19 @@ class Profiles(object):
 
         # write interpolated profiles.yml
         interpoalted_profiles_full_path = interpolated_profiles_dir / PROFILES_FILE_NAME
-        try:
-            with interpoalted_profiles_full_path.open('w') as f:
-                yaml.safe_dump(self.profiles, f)
-        except Exception as e:
-            raise ProfilesError(
-                f'Failed to write interpolated `{PROFILES_FILE_NAME}`' +
-                f' to `{interpolated_profiles_dir}`: {e}'
-            )
+        if os.path.exists(interpoalted_profiles_full_path):
+            try:
+                with interpoalted_profiles_full_path.open('w') as f:
+                    yaml.safe_dump(self.profiles, f)
+            except Exception as e:
+                raise ProfilesError(
+                    f'Failed to write interpolated `{PROFILES_FILE_NAME}`' +
+                    f' to `{interpolated_profiles_dir}`: {e}'
+                )
 
-        self.__interpolated_profiles_dir = str(interpolated_profiles_dir)
-        self.is_interpolated = True
-        return self.__interpolated_profiles_dir
+            self.__interpolated_profiles_dir = str(interpolated_profiles_dir)
+            self.is_interpolated = True
+            return self.__interpolated_profiles_dir
 
     def __del__(self) -> None:
         self.clean()
